Cleaned up the nw/convert folder a bit
