+++ nw/gui/docviewer.py
@@ -17,9 +17,9 @@ from PyQt5.QtCore    import Qt
 from PyQt5.QtWidgets import QTextBrowser
 from PyQt5.QtGui     import QTextOption, QFont, QPalette, QColor
 
-from nw.convert.tokenizer import Tokenizer
-from nw.convert.tohtml    import ToHtml
-from nw.enum              import nwItemType
+from nw.convert.tokenizer   import Tokenizer
+from nw.convert.text.tohtml import ToHtml
+from nw.enum                import nwItemType
 
 logger = logging.getLogger(__name__)
